requirements.txt
aiogram==2.24
python-dotenv
